+++ CHANGELOG.md
@@ -5,16 +5,20 @@
 **🎉 Released:**
 - TBC
 
-**Marker Data** no longer officially supports macOS Sonoma starting with version 2.0.0.
+**Marker Data** now only supports macOS Sequoia 15.7 or later, starting with version 2.0.0.
 
 **🔨 Improvements:**
 - Overhaul Marker Data's GUI to Liquid Glass
 - Updated Marker Data's application icon for macOS Tahoe
+- Verified to run under macOS Tahoe and Final Cut Pro 12
 - [Excel export profile](https://markerdata.theacharya.co/in-action/excel-profile/) now has the ability to automatically embed images
 - Added support and compatibility for FCPXML v1.14 (Final Cut Pro 12)
+- Added support for [Final Cut Pro Creator Studio (Subscription Version)](https://markerdata.theacharya.co/faq/#does-marker-data-work-with-final-cut-pro-creator-studio)
+- Dock Icon's Progress now uses Squircle Style
 - Updated Airtable Module Airlift to version 1.3.3
 - Updated Notion Module CSV2Notion Neo to version 2.1.0
 - Updated Pagemaker Module to version 1.1.1
+- Marker Data's Uninstaller has been rewritten in Swift
 - Updated core dependencies
 - Codebase updates for Xcode 26.3